requirements.txt
beautifulsoup4>=4.12.0
feedparser>=6.0.11
python-dotenv>=1.0.1
requests>=2.31.0
trafilatura>=1.12.0
